scripts/hsytrees/requirements.txt
numpy
google-cloud-storage
flask
geopandas
google-cloud-sql
urllib3